@@ -2,6 +2,11 @@
   "name": "cocktails-site",
   "version": "0.0.0",
   "license": "MIT",
+  "engineStrict": true,
+  "engines": {
+    "node": ">=10.9",
+    "npm": ">=6.4.1"
+  },
   "scripts": {
     "ng": "ng",
     "start": "ng serve --aot --o",
@@ -14,48 +19,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.0",
-    "@angular/cdk": "^7.3.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/material": "^7.3.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/animations": "~8.2.0",
+    "@angular/cdk": "~8.1.2",
+    "@angular/common": "~8.2.0",
+    "@angular/compiler": "~8.2.0",
+    "@angular/core": "~8.2.0",
+    "@angular/forms": "~8.2.0",
+    "@angular/material": "^8.1.2",
+    "@angular/platform-browser": "~8.2.0",
+    "@angular/platform-browser-dynamic": "~8.2.0",
+    "@angular/router": "~8.2.0",
     "bootstrap": "^4.1.3",
-    "core-js": "^2.5.4",
     "googlemaps": "^1.12.0",
     "hammerjs": "^2.0.8",
     "ngx-bootstrap": "^3.0.0-RC.1",
     "ngx-page-scroll": "^5.0.0",
     "popper.js": "^1.14.1",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.1",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular-devkit/build-angular": "~0.802.1",
+    "@angular/cli": "~8.2.1",
+    "@angular/compiler-cli": "~8.2.1",
+    "@angular/language-service": "~8.2.1",
     "@types/googlemaps": "^3.30.16",
-    "@types/jasmine": "~2.8.8",
+    "@types/jasmine": "~3.3.16",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "karma-spec-reporter": "0.0.32",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
